+++ Toris/Assets/Scripts/Player/Player/Inventory/InventoryManager.cs
@@ -54,7 +54,6 @@ namespace OutlandHaven.Inventory
                     }
                 }
             }
-
             EnsureSlotItemStates();
         }